chunkformer/requirements.txt
Pillow
pyyaml
sentencepiece
textgrid
pandas
jiwer
huggingface_hub
colorama
pydub